recovery: updater: Do not set permissions on symlinks
With I6b78ebf3b56b4428331a36cacdf43863e3da5102,
support for adding selabels to symlinks was added.

But if selinux is disabled, this would mean set_perm()
would be called on symlinks (instead of set_metadata()).
set_metadata takes care not to apply permissions on
symlinks but set_perm does not, and if the symlink
links to some critical files (like /proc/mounts),
chmod would abort and OTA upgrade fails.

This patch fixes this behaviour.

Change-Id: I8cff984d15a658c96658ae8d7eb1a81b0644dbaf